feat : 회원기능 만들기 : OAuth + session방식 사용하기

+++ pages/api/post/remove.js
@@ -1,13 +1,24 @@
 import { connectDB } from "@/util/database.js";
 import { ObjectId } from "mongodb";
+import { getServerSession } from "next-auth";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(요청, 응답) {
-  if (요청.method == "GET") {
-    console.log(요청.query._id);
-    let db = (await connectDB).db("board");
-    let result = await db
-      .collection("post")
-      .deleteOne({ _id: new ObjectId(요청.query._id) });
-    응답.status(200).json("삭제완료");
+  let session = await getServerSession(요청, 응답, authOptions);
+  if (session.user.email) {
+    if (요청.method == "GET") {
+      let db = (await connectDB).db("board");
+      let 찾은거 = await db
+        .collection("post")
+        .findOne({ _id: new ObjectId(요청.query._id) });
+      if (찾은거.author == session.user.email) {
+        let result = await db
+          .collection("post")
+          .deleteOne({ _id: new ObjectId(요청.query._id) });
+        응답.status(200).json("삭제완료");
+      } else {
+        return 응답.status(500).json("현재유저와 작성자 불일치");
+      }
+    }
   }
 }
